Creating T2mapping program for python
This is the start of the program to ananlyze T2maps using Python. This program
will hopefully eventually allow for the analysis of some portions of the data
with C-code... this will vastly improve the processing speed. Also, this will
better allow for the transport of this program over to Slicer. This will allow
others to use the T2mapping program but using 3D Slicer.